requirements.txt
flet>=0.10.0
requests>=2.28.0
cryptography>=3.4.8
mutagen>=1.45.1
pillow>=9.0.0
